Refractor UI component and adding test
* Move ChallengeView to dedicated package and split in smaller component
* Add test for DataPool and ImageView
